test(DeleteHistoryRepositoryTest.java) : 삭제 내역 리스트 저장
